apps/backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.1
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
requests==2.32.3
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg[binary]==3.2.1
redis==5.0.8
celery==5.4.0
stripe==10.12.0
slowapi==0.1.9
sentry-sdk[fastapi]==2.13.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.2.0
itsdangerous==2.2.0
pyjwt==2.9.0
orjson==3.10.7
websockets==12.0